Adjust to also include middleware in `GetStepTools<typeof inngest>`
