+++ packages/inngest/src/index.ts
@@ -7,7 +7,13 @@ export {
   type ZodEventSchemas,
 } from "./components/EventSchemas";
 export { Inngest } from "./components/Inngest";
-export type { EventsFromOpts } from "./components/Inngest";
+export type {
+  ClientOptionsFromInngest,
+  EventsFromOpts,
+  GetEvents,
+  GetFunctionInput,
+  GetStepTools,
+} from "./components/Inngest";
 export { InngestCommHandler } from "./components/InngestCommHandler";
 export type { ServeHandler } from "./components/InngestCommHandler";
 export { InngestMiddleware } from "./components/InngestMiddleware";
@@ -33,8 +39,6 @@ export type {
   FailureEventArgs,
   FailureEventPayload,
   FunctionOptions,
-  GetEvents,
-  GetStepTools,
   LogLevel,
   RegisterOptions,
   TimeStr,
